+++ packages/react-components/react-drawer/src/components/DrawerInline/renderDrawerInline.tsx
@@ -1,7 +1,5 @@
-/** @jsxRuntime classic */
-/** @jsx createElement */
-
-import { createElement } from '@fluentui/react-jsx-runtime';
+/** @jsxRuntime automatic */
+/** @jsxImportSource @fluentui/react-jsx-runtime */
 import { assertSlots } from '@fluentui/react-utilities';
 import type { DrawerInlineState, DrawerInlineSlots } from './DrawerInline.types';
 
@@ -9,11 +7,11 @@ import type { DrawerInlineState, DrawerInlineSlots } from './DrawerInline.types'
  * Render the final JSX of DrawerInline
  */
 export const renderDrawerInline_unstable = (state: DrawerInlineState) => {
-  assertSlots<DrawerInlineSlots>(state);
-
-  if (!state.motion.canRender()) {
+  if (!state.motion.canRender) {
     return null;
   }
+
+  assertSlots<DrawerInlineSlots>(state);
 
   return <state.root />;
 };