+++ packages/react-components/react-drawer/src/components/DrawerOverlay/useDrawerOverlay.ts
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { getNativeElementProps, slot, useMergedRefs } from '@fluentui/react-utilities';
 import type { DrawerOverlayProps, DrawerOverlayState } from './DrawerOverlay.types';
 import { DialogProps, DialogSurface, DialogSurfaceProps } from '@fluentui/react-dialog';
-import { useBaseDrawerDefaultProps } from '../../util/useBaseDrawerDefaultProps';
-import { useMotionFromSlot } from '@fluentui/react-motion-preview';
+import { useDrawerDefaultProps } from '../../util/useDrawerDefaultProps';
+import { useMotion } from '@fluentui/react-motion-preview';
 
 /**
  * Create the state required to render DrawerOverlay.
@@ -18,27 +18,26 @@ export const useDrawerOverlay_unstable = (
   props: DrawerOverlayProps,
   ref: React.Ref<HTMLDivElement>,
 ): DrawerOverlayState => {
-  const { open, defaultOpen, size, position } = useBaseDrawerDefaultProps(props);
+  const { open, defaultOpen, size, position } = useDrawerDefaultProps(props);
   const { modalType = 'modal', inertTrapFocus, onOpenChange } = props;
 
-  const [drawerProps, drawerMotion] = useMotionFromSlot(props, open);
-  const [backdropProps, backdropMotion] = useMotionFromSlot(props.backdrop, open);
+  const drawerMotion = useMotion<HTMLDivElement>(open);
+  const backdropMotion = useMotion<HTMLDivElement>(open);
 
-  const hasCustomBackdrop = modalType !== 'non-modal' && backdropProps !== null;
+  const hasCustomBackdrop = modalType !== 'non-modal' && props.backdrop !== null;
 
   const root = slot.always<DialogSurfaceProps>(
     getNativeElementProps('div', {
-      ...drawerProps,
+      ...props,
       ref: useMergedRefs(ref, drawerMotion.ref),
     }),
     {
       elementType: DialogSurface,
       defaultProps: {
-        backdrop: slot.optional(backdropProps, {
+        backdrop: slot.optional(props.backdrop, {
           elementType: 'div',
           renderByDefault: hasCustomBackdrop,
           defaultProps: {
-            motion: backdropMotion,
             ref: backdropMotion.ref,
           },
         }),