+++ packages/react-components/react-drawer/stories/Drawer/DrawerDefault.stories.tsx
@@ -65,7 +65,7 @@ export const Default = () => {
       </Drawer>
 
       <div className={styles.content}>
-        <Button appearance="primary" onClick={onClick}>
+        <Button appearance="primary" onClick={() => setIsOpen(!isOpen)}>
           {type === 'inline' ? 'Toggle' : 'Open'}
         </Button>