+++ packages/react-components/react-tags-preview/stories/TagGroup/TagGroupSizes.stories.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import {
   TagGroup,
-  Tag,
   InteractionTag,
   InteractionTagPrimary,
   InteractionTagSecondary,
@@ -27,12 +26,12 @@ export const Sizes = () => {
         <div key={size}>
           {`${size}: `}
           <TagGroup size={size} aria-label={`${size} tag group example`}>
-            <Tag dismissible media={<Avatar name="Katri Athokas" />}>
-              {size}
-            </Tag>
-            <Tag icon={<CalendarMonthRegular />} shape="circular" dismissible>
-              {size}
-            </Tag>
+            <InteractionTag>
+              <InteractionTagPrimary media={<Avatar name="Katri Athokas" />}>{size}</InteractionTagPrimary>
+            </InteractionTag>
+            <InteractionTag shape="circular">
+              <InteractionTagPrimary icon={<CalendarMonthRegular />}>{size}</InteractionTagPrimary>
+            </InteractionTag>
             <InteractionTag>
               <InteractionTagPrimary icon={<CalendarMonthRegular />} hasSecondaryAction>
                 {size}