+++ packages/react-components/react-toast/src/components/Toaster/renderToaster.tsx
@@ -1,8 +1,6 @@
-/** @jsxRuntime classic */
-/** @jsxFrag Fragment */
-/** @jsx createElement */
+/** @jsxRuntime automatic */
+/** @jsxImportSource @fluentui/react-jsx-runtime */
 
-import { createElement, Fragment } from '@fluentui/react-jsx-runtime';
 import { assertSlots } from '@fluentui/react-utilities';
 import { Portal } from '@fluentui/react-portal';
 import type { ToasterState, ToasterSlotsInternal } from './Toaster.types';
@@ -21,7 +19,7 @@ export const renderToaster_unstable = (state: ToasterState) => {
     <>
       {renderAriaLive ? <AriaLive announceRef={announceRef} /> : null}
       {hasToasts ? (
-        <Portal>
+        <Portal mountNode={state.mountNode}>
           {state.bottomStart ? <state.bottomStart /> : null}
           {state.bottomEnd ? <state.bottomEnd /> : null}
           {state.topStart ? <state.topStart /> : null}